research_agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
openai>=1.40.0
httpx>=0.27.0
ddgs>=9.0.0
# Bundled Scrapling/ (parser + Fetcher): match Scrapling/pyproject base + fetchers
lxml>=6.0.2
cssselect>=1.4.0
orjson>=3.11.7
tld>=0.13.2
w3lib>=2.4.1
typing_extensions
curl_cffi>=0.14.0
# Scrapling MCP + CLI (scrapling[ai] subset; browsers via `scrapling install`)
mcp>=1.26.0
markdownify>=1.2.0
click>=8.3.0
pydantic>=2.0
